# api/src/main/resources/api.yaml
openapi: 3.0.1
info:
  version: '1.0'
  title: Order service
servers:
  - url: '/api/v1/order-service'
...
paths:
...
  /orders/{id}/total:
    get:
      tags:
        - Order Management
      summary: Get order total
      description: Get order total
      operationId: getOrderTotal
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: All existing orders
          content:
            application/json:
              schema:
                type: number
                format: double
...
components:
  parameters:
    id:
      name: id
      in: path
      description: Order (or User) ID
      required: true
      schema:
        type: string
...